Lazy load as many references to the gltf exporter module as possible
